Move validation to status control
